# 7950: Better DisjointBuilder performance for events not ordered by event end attribute