@egahlin 2022-11-09
When I wrote the original code in 2010 collisions were rare. I think saw a few cases of 2-3 lanes (for the same event type and thread), typically when doing recursive events. I wonder what has changed, if this is no longer a case? The JVM usually doesn't write overlapping events.